+++ src/tap/communication/serial/terminal_serial.hpp
@@ -17,8 +17,8 @@
  * along with Taproot.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef TERMINAL_SERIAL_HPP_
-#define TERMINAL_SERIAL_HPP_
+#ifndef TAPROOT_TERMINAL_SERIAL_HPP_
+#define TAPROOT_TERMINAL_SERIAL_HPP_
 
 #include <cstring>
 #include <map>
@@ -169,4 +169,4 @@ private:
 };  // class TerminalSerial
 }  // namespace tap::communication::serial
 
-#endif  // TERMINAL_SERIAL_HPP_
+#endif  // TAPROOT_TERMINAL_SERIAL_HPP_