+++ src/tap/motor/dji_motor_tx_handler.cpp
@@ -30,8 +30,8 @@ namespace tap::motor
 {
 void DjiMotorTxHandler::addMotorToManager(DjiMotor** canMotorStore, DjiMotor* const motor)
 {
-    modm_assert(motor, "DjiMotorTxHandler", "nullptr");
-    int16_t idIndex = DJI_MOTOR_TO_NORMALIZED_ID(motor->getMotorIdentifier());
+    assert(motor != nullptr);
+    uint32_t idIndex = DJI_MOTOR_TO_NORMALIZED_ID(motor->getMotorIdentifier());
     bool motorOverloaded = canMotorStore[idIndex] != nullptr;
     bool motorOutOfBounds = (idIndex < 0) || (idIndex >= DJI_MOTORS_PER_CAN);
     modm_assert(!motorOverloaded && !motorOutOfBounds, "DjiMotorTxHandler", "overloading");