+++ src/tap/motor/dji_motor_tx_handler.hpp
@@ -17,8 +17,8 @@
  * along with Taproot.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef DJI_MOTOR_TX_HANDLER_HPP_
-#define DJI_MOTOR_TX_HANDLER_HPP_
+#ifndef TAPROOT_DJI_MOTOR_TX_HANDLER_HPP_
+#define TAPROOT_DJI_MOTOR_TX_HANDLER_HPP_
 
 #include <limits.h>
 
@@ -120,4 +120,4 @@ private:
 
 }  // namespace tap::motor
 
-#endif  // DJI_MOTOR_TX_HANDLER_HPP_
+#endif  // TAPROOT_DJI_MOTOR_TX_HANDLER_HPP_